Refactor the steeldesign module.
The complicated pickling of the shapes database has been replaced by a
class-based method that reads a compressed CSV file. This simplifies
providing units for the values, and with US and SI values separated
ahead of time, there's no need to develop complicated parsing logic.

The main entry points for users is now the ``ShapesTable`` class, which
provides the methods ``get_prop`` and ``lightest_shape``. ``get_prop``
is a newer, better version of ``property_lookup`` that provides the
value with the units it was defined in! Also, the underlying DataFrame
is now indexed by AISC_Manual_Label, greatly simplifying the code for
finding a shape.

The latex_name function was moved to the `text` module, but I'm seeing
now that I have a tiny module called `latex_format` that perhaps need
some TLC.