Add different way of accessing shape data.
This ended up as an experiment to learn how dynamically created classes
work, but I think it's interesting. I don't think it's any more useful
than just returning a Series from the ShapesTable, but maybe if code
introspection gets *real good* it'll be useful.

For now, it will sit in exile in this branch.