Rework ShapesTable loading (again?)
Adds concept of 'shape resources' -- registered packages of metadata
that describe a ShapesTable, including where it's data sits on disk.
Still loads the tables on import, but not doing so would be a huge
breaking change. Probably. (Maybe the ShapesTable object *itself*
handles the lazy-loading? Ugh, but that gets away from the generality of
it -- theoretically you can gin up a ShapesTable from anywhere,
including if you just write all your data in the Python file.)